pyproject.toml
[tool.poetry]
name = "ai-audiobooks"
version = "0.1.0"
description = "A library for taking any arbitrary file and converting it into an audiobook using AI. Please do not use commercially."
authors = ["Ryan Peach <rgpeach10@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ai_audiobooks"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.14.3"
typer = "^0.10.0"
GitPython = "^3.1.42"
guidance = "^0.1.13"
python-dotenv = "^1.0.1"
pypandoc = "^1.13"
tiktoken = "^0.6.0"
elevenlabs = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
mypy = "^1.9.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
